Remove Encode/Decode helpers and used them in 'Track'
